Refactor: 유저 콘텐츠 컴포넌트 분리
